requirements.txt
pandas>=1.5
pyyaml>=6.0
openpyxl>=3.0
pytest>=7.0
